moved tuple helper functions to newly added stuff/tuple_tools.h
