+++ g2o/core/base_constant_edge.hpp
@@ -48,32 +48,6 @@ private:
   OptimizableGraph::Vertex& _vertex;
 };
 
-template<int I>
-struct Tuple_apply_i
-{
-  template<typename F, typename T>
-  void operator()(F&& f, T& t, int i)
-  {
-    if(i == I - 1)
-      return f(std::get<I - 1>(t));
-    else
-      return Tuple_apply_i<I - 1>()(f, t, i);
-  }
-};
-
-template<>
-struct Tuple_apply_i<0>
-{
-  template<typename F, typename T>
-  void operator()(F&&, T&, int) { }
-};
-
-template<typename F, typename T>
-void tuple_apply_i(F&& f, T& t, int i)
-{
-  Tuple_apply_i<std::tuple_size<T>::value>()(f, t, i);
-}
-
 } // anonymous namespace
 
 template <int D, typename E, typename... VertexTypes>
