+++ g2o/stuff/CMakeLists.txt
@@ -4,7 +4,7 @@ add_library(stuff ${G2O_LIB_TYPE}
   command_args.cpp   misc.h           sparse_helper.cpp  timeutil.h
   command_args.h     os_specific.c    sparse_helper.h
   filesys_tools.cpp  os_specific.h    string_tools.cpp
-  property.cpp       property.h       
+  property.cpp       property.h       tuple_tools.h
   sampler.cpp        sampler.h        unscented.h
   tictoc.cpp tictoc.h
   g2o_stuff_api.h
